Reorganize geometry-related features, add `geometry` component
There are a lot of geometry features in Goost, so it deserves its own component rather than being in `math`. Current documentation also refers to "Geometry" component.

Moved `Random2D` to geometry component, because it relies on it.
