--- core/math/3d/SCsub
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-
-# Nothing here yet! 
-# Are you the one who's going to implement this? :)
