+++ core/math/geometry/2d/SCsub
