+++ core/math/geometry/2d/goost_geometry_2d.cpp
