+++ core/math/geometry/2d/poly/SCsub
