+++ core/math/geometry/2d/poly/boolean/clipper10/poly_boolean_clipper10.h
