+++ core/math/geometry/2d/poly/boolean/clipper6/poly_boolean_clipper6.h
