+++ core/math/geometry/2d/poly/boolean/poly_boolean.cpp
