+++ core/math/geometry/2d/poly/decomp/clipper10/poly_decomp_clipper10.cpp
@@ -1,5 +1,5 @@
 #include "poly_decomp_clipper10.h"
-#include "goost/core/math/2d/geometry/poly/utils/godot_clipper10_path_convert.h"
+#include "goost/core/math/geometry/2d/poly/utils/godot_clipper10_path_convert.h"
 
 Vector<Vector<Point2>> PolyDecomp2DClipper10::triangulate_mono(const Vector<Vector<Point2>> &p_polygons) {
 	using namespace clipperlib;
