+++ core/math/geometry/2d/poly/offset/clipper10/poly_offset_clipper10.h
