+++ core/math/geometry/2d/poly/offset/clipper6/poly_offset_clipper6.h
