+++ core/math/geometry/2d/poly/poly_node_2d.h
