+++ core/math/geometry/2d/poly/utils/godot_clipper10_path_convert.h
