+++ core/math/geometry/2d/poly/utils/godot_clipper6_path_convert.h
