+++ core/math/geometry/register_geometry_types.h
@@ -1,0 +1,6 @@
+namespace goost {
+
+void register_geometry_types();
+void unregister_geometry_types();
+
+} // namespace goost
