+++ register_types.h
@@ -20,13 +20,13 @@
 #include "core/image/image_blender.h"
 #include "core/image/image_indexed.h"
 #include "core/invoke_state.h"
-#include "core/math/2d/geometry/goost_geometry_2d.h"
-#include "core/math/2d/geometry/goost_geometry_2d_bind.h"
-#include "core/math/2d/geometry/poly/boolean/poly_boolean.h"
-#include "core/math/2d/geometry/poly/decomp/poly_decomp.h"
-#include "core/math/2d/geometry/poly/offset/poly_offset.h"
-#include "core/math/2d/geometry/poly/poly_backends.h"
-#include "core/math/2d/random_2d.h"
+#include "core/math/geometry/2d/goost_geometry_2d.h"
+#include "core/math/geometry/2d/goost_geometry_2d_bind.h"
+#include "core/math/geometry/2d/poly/boolean/poly_boolean.h"
+#include "core/math/geometry/2d/poly/decomp/poly_decomp.h"
+#include "core/math/geometry/2d/poly/offset/poly_offset.h"
+#include "core/math/geometry/2d/poly/poly_backends.h"
+#include "core/math/geometry/2d/random_2d.h"
 #include "core/math/random.h"
 #include "core/types/linked_list.h"
 #include "core/types/variant_map.h"
